Made hooks use api.js and removed useFetch hook
